# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cottage Pricing API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: settings
  - name: calendar
  - name: bookings
  - name: history
  - name: rentals
  - name: auth
  - name: cottage
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all control panel settings
      responses:
        "200":
          description: Settings object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update control panel settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

  /calendar:
    get:
      operationId: getCalendar
      tags: [calendar]
      summary: Get calendar with computed daily prices
      parameters:
        - name: fromYear
          in: query
          required: false
          schema:
            type: integer
          description: First year to include (defaults to current year)
        - name: toYear
          in: query
          required: false
          schema:
            type: integer
          description: Last year to include (defaults to fromYear + 1)
      responses:
        "200":
          description: List of calendar entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEntry"

  /calendar/bulk-days:
    post:
      operationId: setBulkDays
      tags: [calendar]
      summary: Apply sequential day-of-week overrides starting from a given day for all calendar entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDaysRequest"
      responses:
        "200":
          description: Number of entries updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDaysResult"

  /calendar/{date}/override:
    put:
      operationId: setDayOverride
      tags: [calendar]
      summary: Set per-day overrides for season, holiday, or day-of-week
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DayOverrideRequest"
      responses:
        "200":
          description: Updated calendar entry for that date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEntry"
    delete:
      operationId: removeDayOverride
      tags: [calendar]
      summary: Remove per-day override
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Calendar entry with override removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEntry"

  /bookings/calculate:
    post:
      operationId: calculateBooking
      tags: [bookings]
      summary: Calculate total price and nightly breakdown for a date range
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingRequest"
      responses:
        "200":
          description: Booking calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResult"

  /rentals:
    get:
      operationId: getRentals
      tags: [rentals]
      summary: List all rentals
      responses:
        "200":
          description: List of rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RentalEntry"
    post:
      operationId: createRental
      tags: [rentals]
      summary: Create a new rental booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentalInput"
      responses:
        "201":
          description: Created rental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalEntry"

  /rentals/{id}:
    get:
      operationId: getRental
      tags: [rentals]
      summary: Get a single rental
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rental detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalEntry"
    patch:
      operationId: updateRental
      tags: [rentals]
      summary: Update a rental (status, details, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentalPatch"
      responses:
        "200":
          description: Updated rental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalEntry"
    delete:
      operationId: deleteRental
      tags: [rentals]
      summary: Delete a rental
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                required: [deleted]

  /rentals/{id}/approvals:
    get:
      operationId: getRentalApprovals
      tags: [rentals]
      summary: Get approval records for a rental
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OwnerApproval"

  /rentals/{id}/approvals/{ownerEmail}:
    patch:
      operationId: setRentalApproval
      tags: [rentals]
      summary: Approve or un-approve a rental on behalf of an owner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: ownerEmail
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalPatch"
      responses:
        "200":
          description: Updated rental with approvals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalEntry"

  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: JWT token + user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"

  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [auth]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{id}:
    patch:
      operationId: updateUser
      tags: [auth]
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatch"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [auth]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                required: [deleted]

  /cottage-info:
    get:
      operationId: getCottageInfo
      tags: [cottage]
      summary: Get cottage information (title, description, photos)
      responses:
        "200":
          description: Cottage info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CottageInfo"
    put:
      operationId: updateCottageInfo
      tags: [cottage]
      summary: Update cottage info (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CottageInfo"
      responses:
        "200":
          description: Updated cottage info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CottageInfo"

  /history:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get change history log
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
          description: Max number of entries to return (default 200)
      responses:
        "200":
          description: List of change history entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChangeHistoryEntry"
    delete:
      operationId: clearHistory
      tags: [history]
      summary: Clear all change history
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                required: [deleted]

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Season:
      type: object
      properties:
        name:
          type: string
        multiplier:
          type: number
        startDate:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
      required: [name, multiplier]

    Holiday:
      type: object
      properties:
        name:
          type: string
        boost:
          type: number
        startDate:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
      required: [name, boost]

    Owner:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
      required: [name, email]

    DayMultipliers:
      type: object
      properties:
        Monday: { type: number }
        Tuesday: { type: number }
        Wednesday: { type: number }
        Thursday: { type: number }
        Friday: { type: number }
        Saturday: { type: number }
        Sunday: { type: number }
      required: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]

    Settings:
      type: object
      properties:
        basePrice:
          type: number
        familyRate:
          type: number
        seasons:
          type: array
          items:
            $ref: "#/components/schemas/Season"
        dayMultipliers:
          $ref: "#/components/schemas/DayMultipliers"
        holidays:
          type: array
          items:
            $ref: "#/components/schemas/Holiday"
        holidaysByYear:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/Holiday"
          description: "Per-year holiday overrides keyed by year string (e.g. '2026')"
        owners:
          type: array
          items:
            $ref: "#/components/schemas/Owner"
      required: [basePrice, familyRate, seasons, dayMultipliers, holidays]

    DayOverrideRequest:
      type: object
      properties:
        seasonOverride:
          type: string
          nullable: true
        holidayOverride:
          type: string
          nullable: true
        dayOverride:
          type: string
          nullable: true

    BulkDaysRequest:
      type: object
      properties:
        startDay:
          type: string
          description: "Day name to assign to the first calendar entry (e.g. Monday)"
        fromYear:
          type: integer
          nullable: true
        toYear:
          type: integer
          nullable: true
      required: [startDay]

    BulkDaysResult:
      type: object
      properties:
        updated:
          type: integer
      required: [updated]

    CalendarEntry:
      type: object
      properties:
        date:
          type: string
        dayOfWeek:
          type: string
        season:
          type: string
        seasonMult:
          type: number
        dayMult:
          type: number
        holiday:
          type: string
        holidayBoost:
          type: number
        demandAdj:
          type: number
        finalPct:
          type: number
        price:
          type: number
        isOverridden:
          type: boolean
        seasonIsOverridden:
          type: boolean
        holidayIsOverridden:
          type: boolean
        dayIsOverridden:
          type: boolean
        syncedSeason:
          type: string
          nullable: true
        syncedHoliday:
          type: string
          nullable: true
      required: [date, dayOfWeek, season, seasonMult, dayMult, holiday, holidayBoost, demandAdj, finalPct, price, isOverridden, seasonIsOverridden, holidayIsOverridden, dayIsOverridden]

    BookingRequest:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        rateType:
          type: string
          nullable: true
        includeMultipliers:
          type: boolean
          nullable: true
      required: [startDate, endDate]

    BookingResult:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        nights:
          type: integer
        totalPrice:
          type: number
        avgDailyRate:
          type: number
        rateType:
          type: string
        includeMultipliers:
          type: boolean
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/CalendarEntry"
      required: [startDate, endDate, nights, totalPrice, avgDailyRate, rateType, includeMultipliers, breakdown]

    RentalEntry:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
        renterName:
          type: string
        phone:
          type: string
        email:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        nights:
          type: integer
        totalPrice:
          type: number
        agreedPrice:
          type: number
          nullable: true
        rateType:
          type: string
        bookingType:
          type: string
        extraDetails:
          type: string
        status:
          type: string
      required: [id, createdAt, renterName, phone, email, startDate, endDate, nights, totalPrice, rateType, bookingType, extraDetails, status]

    RentalInput:
      type: object
      properties:
        renterName:
          type: string
        phone:
          type: string
        email:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        nights:
          type: integer
        totalPrice:
          type: number
        rateType:
          type: string
        bookingType:
          type: string
        extraDetails:
          type: string
      required: [renterName, phone, email, startDate, endDate, nights, totalPrice, rateType]

    RentalPatch:
      type: object
      properties:
        status:
          type: string
          nullable: true
        renterName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        extraDetails:
          type: string
          nullable: true
        agreedPrice:
          type: number
          nullable: true
        sendOwnerEmail:
          type: boolean
          nullable: true
        sendRenterEmail:
          type: boolean
          nullable: true

    OwnerApproval:
      type: object
      properties:
        id:
          type: integer
        rentalId:
          type: integer
        ownerEmail:
          type: string
        ownerName:
          type: string
        approved:
          type: boolean
        approvedAt:
          type: string
          nullable: true
      required: [id, rentalId, ownerEmail, ownerName, approved]

    ApprovalPatch:
      type: object
      properties:
        approved:
          type: boolean
      required: [approved]

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required: [id, email, name, role, createdAt]

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    UserInput:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
      required: [email, name, password, role]

    UserPatch:
      type: object
      properties:
        name:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        password:
          type: string
          nullable: true

    CottageInfo:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        photos:
          type: array
          items:
            type: string
      required: [title, description, photos]

    ChangeHistoryEntry:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
        changeType:
          type: string
        description:
          type: string
        metadata:
          type: string
      required: [id, createdAt, changeType, description, metadata]