# lib/api-zod/src/generated/types/cottageInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cottage Pricing API
 * OpenAPI spec version: 0.1.0
 */

export interface CottageInfo {
  title: string;
  description: string;
  photos: string[];
}